+++ apps/web-svelte/src/routes/components-demo/+page.svelte
@@ -1,0 +1,5 @@
+<script lang="ts">
+  import ComponentsDemo from './components-demo.svelte'
+</script>
+
+<ComponentsDemo />